NamedGraphs

Documentation for NamedGraphs.

NamedGraphs.DefaultNamedCapacity

NamedGraphs.Key

NamedGraphs.NamedDijkstraState

NamedGraphs.add_edges!

NamedGraphs.incident_edges

NamedGraphs.is_path_graph

NamedGraphs.rem_edges!

NamedGraphs.DefaultNamedCapacity — Type

DefaultNamedCapacity{T}

Structure that returns 1 if a forward edge exists in flow_graph, and 0 otherwise.

NamedGraphs.Key — Type

Key{K}

A key (index) type, used for unambiguously identifying an object as a key or index of an indexible object AbstractArray, AbstractDict, etc.

Useful for nested structures of indices, for example:

[Key([1, 2]), [Key([3, 4]), Key([5, 6])]]

which could represent partitioning a set of vertices

[Key([1, 2]), Key([3, 4]), Key([5, 6])]

NamedGraphs.NamedDijkstraState — Type

struct NamedDijkstraState{V,T}

An AbstractPathState designed for Dijkstra shortest-paths calculations.

NamedGraphs.add_edges! — Method

Add a list of edges to a graph g

NamedGraphs.incident_edges — Method

incident_edges(graph::AbstractGraph, vertex; dir=:out)

Edges incident to the vertex vertex.

dir ∈ (:in, :out, :both), defaults to :out.

For undirected graphs, returns all incident edges.

Like: https://juliagraphs.org/Graphs.jl/v1.7/algorithms/linalg/#Graphs.LinAlg.adjacency_matrix

NamedGraphs.is_path_graph — Method

TODO: Make this more sophisticated, check that only two vertices have degree 1 and none have degree 0, meaning it is a path/linear graph:

https://en.wikipedia.org/wiki/Path_graph

but not a path/linear forest:

https://en.wikipedia.org/wiki/Linear_forest

NamedGraphs.rem_edges! — Method

Remove a list of edges from a graph g