Reference

NamedGraphs.DefaultNamedCapacity — Type

DefaultNamedCapacity{T}

Structure that returns 1 if a forward edge exists in flow_graph, and 0 otherwise.

NamedGraphs.NamedDijkstraState — Type

struct NamedDijkstraState{V,T}

An AbstractPathState designed for Dijkstra shortest-paths calculations.

NamedGraphs.edgeinduced_subgraphs_no_leaves — Method

edgeinduced_subgraphs_no_leaves(g::AbstractNamedGraph, max_number_of_edges::Int64)

Enumerate all unique, connected edgesubgraphs without any leaf vertices (degree 1) and with Nedges <= maxnumberof_edges

NamedGraphs.GraphsExtensions.CURRENT_PARTITIONING_BACKEND — Constant

Current default graph partitioning backend

NamedGraphs.GraphsExtensions.Backend — Type

Graph partitioning backend

NamedGraphs.GraphsExtensions.add_edges! — Method

Add a list of edges to a graph g

NamedGraphs.GraphsExtensions.current_partitioning_backend — Method

Get the graph partitioning backend

NamedGraphs.GraphsExtensions.distance_to_leaves — Method

Get distance of a vertex from a leaf

NamedGraphs.GraphsExtensions.has_leaf_neighbor — Method

Determine if a node has any neighbors which are leaves

NamedGraphs.GraphsExtensions.incident_edges — Method

incident_edges(graph::AbstractGraph, vertex; dir=:out)

Edges incident to the vertex vertex.

dir ∈ (:in, :out, :both), defaults to :out.

For undirected graphs, returns all incident edges.

Like: https://juliagraphs.org/Graphs.jl/v1.7/algorithms/linalg/#Graphs.LinAlg.adjacency_matrix

NamedGraphs.GraphsExtensions.is_cycle_graph — Method

https://juliagraphs.org/Graphs.jl/dev/corefunctions/simplegraphsgenerators/#Graphs.SimpleGraphs.cyclegraph-Tuple%7BT%7D%20where%20T%3C:Integer https://en.wikipedia.org/wiki/Cyclegraph

NamedGraphs.GraphsExtensions.is_leaf_edge — Method

Determine if an edge involves a leaf (at src or dst)

NamedGraphs.GraphsExtensions.is_path_graph — Method

Check if an undirected graph is a path/linear graph:

https://en.wikipedia.org/wiki/Path_graph

but not a path/linear forest:

https://en.wikipedia.org/wiki/Linear_forest

NamedGraphs.GraphsExtensions.non_leaf_edges — Method

Get all edges which do not involve a leaf

https://en.wikipedia.org/wiki/Tree(graphtheory)#Definitions

NamedGraphs.GraphsExtensions.random_bfs_tree — Method

Do a BFS search to construct a tree, but do it with randomness to avoid generating the same tree. Based on Int. J. Comput. Their Appl. 15 pp 177-186 (2008). Edges will point away from source vertex s.

NamedGraphs.GraphsExtensions.rem_edges! — Method

Remove a list of edges from a graph g

NamedGraphs.GraphsExtensions.rem_vertices! — Method

Remove a list of vertices from a graph g

NamedGraphs.GraphsExtensions.root_vertex — Method

Return the root vertex of a rooted directed graph.

This will return the first root vertex that is found, so won't error if there is more than one.

NamedGraphs.GraphsExtensions.set_partitioning_backend! — Method

Set the graph partitioning backend

NamedGraphs.NamedGraphGenerators.named_hexagonal_lattice_graph — Method

Generate a graph which corresponds to a hexagonal tiling of the plane. There are m rows and n columns of hexagons. Based off of the generator in Networkx hexagonallatticegraph()

NamedGraphs.NamedGraphGenerators.named_triangular_lattice_graph — Method

Generate a graph which corresponds to a equilateral triangle tiling of the plane. There are m rows and n columns of triangles. Based off of the generator in Networkx triangularlatticegraph()

NamedGraphs.Keys.Key — Type

Key{K}

A key (index) type, used for unambiguously identifying an object as a key or index of an indexible object AbstractArray, AbstractDict, etc.

Useful for nested structures of indices, for example:

[Key([1, 2]), [Key([3, 4]), Key([5, 6])]]

which could represent partitioning a set of vertices

[Key([1, 2]), Key([3, 4]), Key([5, 6])]